admin panel

+++ database/migrations/2025_08_25_135538_add_role_to_users_table.php
@@ -9,14 +9,19 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->string('role')->default('user'); // default user
+            if (!Schema::hasColumn('users', 'role')) {
+                // default staff biar jelas bedain dengan admin
+                $table->string('role', 20)->default('staff')->after('password');
+            }
         });
     }
 
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('role');
+            if (Schema::hasColumn('users', 'role')) {
+                $table->dropColumn('role');
+            }
         });
     }
 };
